Removed unecessary cast

+++ src/managers/SettingsManager.cpp
@@ -8,7 +8,7 @@ CustomStruct::DisplaySettings SettingsManager::Display = {
     Mod::get()->getSettingValue<int>("text-width-offset"),
     Mod::get()->getSettingValue<int>("text-height-offset"),
     Mod::get()->getSettingValue<float>("text-size"),
-    static_cast<int>(round(Mod::get()->getSettingValue<int>("text-opacity") / 100.f * 255)),
+    round(Mod::get()->getSettingValue<int>("text-opacity") / 100.f * 255),
     Mod::get()->getSettingValue<ccColor3B>("text-color"),
     Mod::get()->getSettingValue<std::string>("number-separator")
 };
@@ -52,7 +52,7 @@ $execute {
         SettingsManager::Display.size = size;
     });
     listenForSettingChanges<int>("text-opacity", [](int opacity) {
-        SettingsManager::Display.opacity = static_cast<int>(round(opacity / 100.f * 255));
+        SettingsManager::Display.opacity = round(opacity / 100.f * 255);
     });
     listenForSettingChanges<ccColor3B>("text-color", [](ccColor3B color) {
         SettingsManager::Display.color = color;
